# core-banking-psdx-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-psdx
  description: PSD2/PSD3 and FIDA Regulatory Compliance Service
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/psdx
    description: Development Environment
  - url: /
    description: Local Development Environment
security:
  - apiKey: []
  - bearer: []
...
paths:
...
  /api/v1/access-logs:
    get:
      tags:
        - Access Logging
      summary: Get access logs
      description: Gets access logs based on query parameters
      operationId: getAccessLogs
      parameters:
        - name: partyId
          in: query
          description: ID of the customer
          required: false
          schema:
            type: string
            format: uuid
        - name: consentId
          in: query
          description: ID of the consent
          required: false
          schema:
            type: string
            format: uuid
        - name: thirdPartyId
          in: query
          description: ID of the third party provider
          required: false
          schema:
            type: string
        - name: fromDate
          in: query
          description: Start date of the range
          required: false
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          description: End date of the range
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Consent-ID
          in: header
          description: ID of the consent for the operation
          schema:
            type: string
          example: '12345'
      responses:
        '200':
          description: Access logs found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PSDAccessLogDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PSDAccessLogDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PSDAccessLogDTO'
    post:
      tags:
        - Access Logging
      summary: Log access
      description: Logs an access to customer data or an operation performed on their behalf
      operationId: logAccess
      parameters:
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Consent-ID
          in: header
          description: ID of the consent for the operation
          schema:
            type: string
          example: '12345'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PSDAccessLogRequestDTO'
        required: true
      responses:
        '201':
          description: Access logged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAccessLogDTO'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAccessLogDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAccessLogDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAccessLogDTO'
...
  /api/v1/access-logs/{logId}:
    get:
      tags:
        - Access Logging
      summary: Get access log
      description: Gets an access log by its ID
      operationId: getAccessLog
      parameters:
        - name: logId
          in: path
          description: ID of the access log
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Consent-ID
          in: header
          description: ID of the consent for the operation
          schema:
            type: string
          example: '12345'
      responses:
        '200':
          description: Access log found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAccessLogDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAccessLogDTO'
        '404':
          description: Access log not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAccessLogDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAccessLogDTO'
components:
  schemas:
...
    PSDAccessLogRequestDTO:
      required:
        - accessType
        - consentId
        - partyId
        - resourceId
        - resourceType
        - status
        - thirdPartyId
      type: object
      properties:
        consentId:
          type: string
          description: ID of the consent used for the access
          format: uuid
        partyId:
          type: string
          description: ID of the customer
          format: uuid
        thirdPartyId:
          maxLength: 100
          minLength: 0
          pattern: ^[A-Za-z0-9\-_]+$
          type: string
          description: ID of the third party provider
        accessType:
          pattern: ^(READ|WRITE|DELETE)$
          type: string
          description: Type of access
          example: READ
        resourceType:
          pattern: ^(ACCOUNT|PAYMENT|CONSENT|BALANCE|TRANSACTION|FUNDS_CONFIRMATION)$
          type: string
          description: Type of resource being accessed
          example: ACCOUNT
        resourceId:
          maxLength: 100
          minLength: 0
          pattern: ^[A-Za-z0-9\-_]+$
          type: string
          description: ID of the resource being accessed
          example: ACC123456
        ipAddress:
          pattern: ^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$|^(?:[0-9a-fA-F]{1,4}:){7}[0-9a-fA-F]{1,4}$
          type: string
          description: IP address of the requester
          example: 192.168.1.1
        userAgent:
          type: string
          description: User agent of the requester
          example: Mozilla/5.0
        status:
          type: string
          description: Status of the access
          example: SUCCESS
        errorMessage:
          type: string
          description: Error message if the access failed
          example: Unauthorized access
        tppRequestId:
          type: string
          description: TPP-Request-ID header value
          example: TPP-123456
        psuId:
          type: string
          description: PSU-ID header value
          example: PSU-123456
        psuIdType:
          type: string
          description: PSU-ID-Type header value
          example: email
        psuCorporateId:
          type: string
          description: PSU-Corporate-ID header value
          example: CORP-123456
        psuCorporateIdType:
          type: string
          description: PSU-Corporate-ID-Type header value
          example: email
        tppRedirectUri:
          type: string
          description: TPP-Redirect-URI header value
          example: https://tpp.com/callback
        xrequestId:
          type: string
      description: PSD Access log request
    PSDAccessLogDTO:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the access log
          format: uuid
        consentId:
          type: string
          description: ID of the consent used for the access
          format: uuid
        partyId:
          type: string
          description: ID of the customer
          format: uuid
        thirdPartyId:
          type: string
          description: ID of the third party provider
        accessType:
          type: string
          description: Type of access
          example: READ
        resourceType:
          type: string
          description: Type of resource being accessed
          example: ACCOUNT
        resourceId:
          type: string
          description: ID of the resource being accessed
          example: ACC123456
        ipAddress:
          type: string
          description: IP address of the requester
          example: 192.168.1.1
        userAgent:
          type: string
          description: User agent of the requester
          example: Mozilla/5.0
        status:
          type: string
          description: Status of the access
          example: SUCCESS
        errorMessage:
          type: string
          description: Error message if the access failed
          example: Unauthorized access
        tppRequestId:
          type: string
          description: TPP-Request-ID header value
          example: TPP-123456
        psuId:
          type: string
          description: PSU-ID header value
          example: PSU-123456
        psuIdType:
          type: string
          description: PSU-ID-Type header value
          example: email
        psuCorporateId:
          type: string
          description: PSU-Corporate-ID header value
          example: CORP-123456
        psuCorporateIdType:
          type: string
          description: PSU-Corporate-ID-Type header value
          example: email
        tppRedirectUri:
          type: string
          description: TPP-Redirect-URI header value
          example: https://tpp.com/callback
        createdAt:
          type: string
          description: Date and time when the access was logged
          format: date-time
        xrequestId:
          type: string
      description: PSD Access log information
...
  securitySchemes:
    apiKey:
      type: apiKey
      description: API key for TPP authentication
      name: X-API-KEY
      in: header
    bearer:
      type: http
      description: JWT token for user authentication
      scheme: bearer
      bearerFormat: JWT